clean html; remove CRModal (only used in page and small)
